Add HTTP Keep-Alive support (rudimentary, but functional).
This is done by adding two new hooks:
open_req - Opens a socket (send_req used to do this implicitly).
create_req - Constructs the request (used to be part of get_next_url).

The open_req is executed *before* the create_req because we need to know
if we are a keepalive request and only the socket code knows whether we
are keepalive or not.  This is a slight performance penalty on non-keepalive
requests because we are creating the request buffer *after* we have opened
the socket.  We could do a "create_sock" that does everything but connect
and sets the keepalive flag?  I'll discuss this with Aaron on Monday.

We now have a field in the request that specifies if it should be
Keep-Alive or not.  Also, the response has a keepalive field to indicate
whether the connection was keptalive.  If the connection was not keptalive,
then the socket should be closed at each iteration.


git-svn-id: https://svn.apache.org/repos/asf/httpd/test/trunk/flood@89764 13f79535-47bb-0310-9956-ffa450edef68