Merge SSL socket code in with the normal socket calls.  Now determined based on the scheme.
Note that the flood_net.c code may return APR_EGENERAL (there really is
no way to map to APR_EOF or APR_ETIMEUP with OpenSSL's API), so that
must also be checked and treated as an error.

Remove the flood_socket_ssl.c code.  We were doing the abstraction twice.
Once at the flood_socket level and the flood_net level, this is cleaner.


git-svn-id: https://svn.apache.org/repos/asf/httpd/test/trunk/flood@89952 13f79535-47bb-0310-9956-ffa450edef68