Added a new parameter --enable-ssl. SSL is enabled by default, so if you don't have the libs you either have to install them or pass --disable-ssl.
Revamp of the --with-openssl arg to work with the new --enable-ssl arg.
It now just adds the paths to the appropriate compiler/linker flags
and moves on. I left in the -R flag for solaris, but I'm wondering if
I forgot a crossplatform way to do this in libtool. Oh well.

If SSL is enabled (ie not explicitly disabled) it must pass a number of
criteria: existence of various headers, ability to link libssl and
libcrypto (either from given env vars or appropriate --with-openssl
parameters), and the openssl version must be at least 0.9.6. This last
requirement may be further raised to 0.9.6b if the rumors of lower
version crashing under higher load is true.


git-svn-id: https://svn.apache.org/repos/asf/httpd/test/trunk/flood@91991 13f79535-47bb-0310-9956-ffa450edef68