New build system that allows complete separation from httpd-2.0 tree.
Use the find_apr.m4 and find_apu.m4 macros to find and link against
APR/apr-util.

This provides --with-apr= and --with-apr-util= to use an installed
version.  Or, if you place apr and apr-util under the flood tree,
configure will use/build them automatically.  (This is SVN's method
and I rather like it for distribution.)

The compilation system (spec. rules.mk.in) is from httpd-2.0.

There may be some kinks, but it builds here, so let's checkpoint.


git-svn-id: https://svn.apache.org/repos/asf/httpd/test/trunk/flood@92877 13f79535-47bb-0310-9956-ffa450edef68