(Win32) Fixed link error apr_pstrmemdup in flood_socket_keepalive.c.
This patch addresses a link error.  The code invoked apr_pstrmemdup()
in flood_socket_keepalive.c without including the proper declaration.
The compiler generates its own guess, which is incorrect.  The code now
includes the file:
  apr_strings.h
so that it picks up the proper declaration of apr_pstrmemdup.

Below are examples of the compile-time warning and the
link failures that this patch addresses.

  flood_socket_keepalive.c(407) : warning C4013: 'apr_pstrmemdup'
  undefined; assuming extern returning int
  flood_socket_keepalive.obj : error LNK2001: unresolved external
  symbol _apr_pstrmemdup

Submitted by: Phi-Long Tran <ptran@pobox.com>
Reviewed by: Justin Erenkrantz


git-svn-id: https://svn.apache.org/repos/asf/httpd/test/trunk/flood@98529 13f79535-47bb-0310-9956-ffa450edef68