Added generic_fullresp_recv_resp() to force full response.
This patch adds a new handler that always retrieves the entire response.
This handler is needed because we currently do not have a way to tell
flood to always retrieve the entire response.  We have a "wantresponse"
flag that tells the socket code whether or not to retrieve the entire
response.  Handlers such as round_robin_create_req() set this flag based
on varying factors.  With the new generic_fullresp_recv_resp() handler,
you can specify it in your <profile> section like this:
  <recv_resp>generic_fullresp_recv_resp</recv_resp>
to cause flood to retrieve the entire response.

When used with the handler:
  relative_times_process_stats
you can measure the total bytes sent and received with each request
and response.

Submitted by: Phi-Long Tran <ptran@pobox.com>
Reviewed by: Justin Erenkrantz


git-svn-id: https://svn.apache.org/repos/asf/httpd/test/trunk/flood@98633 13f79535-47bb-0310-9956-ffa450edef68